Changes to resolve the discrepancy in GreenButton and YellowRouteButton components
