+++ webapp/src/components/GreenButton/GreenButton.js
@@ -1,15 +1,18 @@
 import React from "react";
 import { Button } from "@material-ui/core";
 import useStyles from "./GreenButtonStyles";
+import { CustomRouterLink } from "../../components";
 
 const GreenButton = props => {
   const classes = useStyles();
   return (
     <Button
-      type={props.type ? props.type : "submit"}
-      color={props.color ? props.color : "primary"}
       variant={props.variant ? props.variant : "contained"}
-      className={classes.submitbtn}
+      color={props.color ? props.color : "primary"}
+      className={classes.routebtn}
+      startIcon={props.startIcon}
+      component={CustomRouterLink}
+      to={props.to}
       onClick={props.onClick}
     >
       {props.children}
