+++ webapp/src/components/GreenButton/GreenButtonStyles.js
@@ -1,16 +1,17 @@
 import { makeStyles } from "@material-ui/core/styles";
 
 const useStyles = makeStyles(theme => ({
-  submitbtn: {
-    backgroundColor: "#f6c80a",
-    color: "#353535",
+  routebtn: {
+    backgroundColor: "#43a047",
+    color: "#fff",
     borderRadius: "3px",
     padding: "0px 12px",
-    textTransform: "Capitalize",
+    textTransform: "capitalize",
     fontSize: "13px",
     fontWeight: "700",
     "&:hover": {
-      background: "#f6c80a"
+      background: "#000",
+      color: "#fff"
     }
   }
 }));
