--- webapp/src/components/YellowRouteButton/index.js
@@ -1,1 +1,0 @@
-export { default } from "./YellowRouteButton";
