+++ webapp/src/components/index.js
@@ -3,9 +3,9 @@ export { default as Auth } from "./Auth";
 export { default as NotFoundPage } from "./NotFoundPage";
 export { default as Validation } from "./Validation";
 export { default as Backdrop } from "./Backdrop";
-export { default as GreenButton } from "../components/GreenButton";
+export { default as YellowButton } from "../components/YellowButton";
 export { default as GrayButton } from "../components/GrayButton";
-export { default as YellowRouteButton } from "../components/YellowRouteButton";
+export { default as GreenButton } from "../components/GreenButton";
 export { default as Input } from "../components/Input";
 export { default as InputText } from "../components/InputText";
 export { default as Logo } from "./Logo";
