+++ webapp/src/containers/RPC/ViewRpc/DeleteRpc.js
@@ -15,7 +15,7 @@ import * as strapiUtilities from "../../../Utilities/StrapiUtilities";
 import * as genericConstants from "../../../constants/GenericConstants";
 // import ZoneSchema from "../ZoneSchema";
 import { get } from "lodash";
-import { GreenButton } from "../../../components";
+import { YellowButton } from "../../../components";
 const RPC_URL = strapiConstants.STRAPI_DB_URL + strapiConstants.STRAPI_RPCS;
 // const ZONE_ID = "stateName";
 
@@ -101,14 +101,14 @@ const DeleteRpc = props => {
                   Do yo want to delete this field?
                 </Grid>
                 <Grid item xs>
-                  <GreenButton
+                  <YellowButton
                     type="submit"
                     color="primary"
                     variant="contained"
                     onClick={handleSubmit}
                   >
                     {genericConstants.DELETE_TEXT}
-                  </GreenButton>
+                  </YellowButton>
                 </Grid>
               </Grid>
             </Grid>
