+++ webapp/src/containers/State/ViewState/DeleteState.js
@@ -8,7 +8,7 @@ import * as strapiConstants from "../../../constants/StrapiApiConstants";
 import useStyles from "./DeleteStateStyles";
 import * as serviceProviders from "../../../api/Axios";
 import * as genericConstants from "../../../constants/GenericConstants";
-import { GreenButton } from "../../../components";
+import { YellowButton } from "../../../components";
 
 const STATE_URL = strapiConstants.STRAPI_DB_URL + strapiConstants.STRAPI_STATES;
 const STATE_ID = "state";
@@ -88,24 +88,25 @@ const DeleteState = props => {
             {genericConstants.DELETE_TEXT}
           </Typography>
           <div className={classes.edit_dialog}>
-          <Grid item xs={12}>
-            <Grid container spacing={2} alignItems="center">
-              <Grid item lg className={classes.deletemessage}>
-                Do yo want to delete this field?
-              </Grid>
-              <Grid item xs>
-                <GreenButton
-                  type="submit"
-                  color="primary"
-                  variant="contained"
-                  onClick={handleSubmit}
-                >
-                  {genericConstants.DELETE_TEXT}
-                </GreenButton>
+            <Grid item xs={12}>
+              <Grid container spacing={2} alignItems="center">
+                <Grid item lg className={classes.deletemessage}>
+                  Do yo want to delete this field?
+                </Grid>
+                <Grid item xs>
+                  <YellowButton
+                    type="submit"
+                    color="primary"
+                    variant="contained"
+                    onClick={handleSubmit}
+                  >
+                    {genericConstants.DELETE_TEXT}
+                  </YellowButton>
+                </Grid>
               </Grid>
             </Grid>
-          </Grid>
-        </div></div>
+          </div>
+        </div>
       </Fade>
     </Modal>
   );
